+++ SpecView/IntensityView.cpp
@@ -29,7 +29,7 @@ using namespace Lexi;
 using namespace Root;
 
 //////////////////////////////////////////////////////////////////////
-static const long ColorMax = Root::ByteMax;
+static const int ColorMax = Root::ByteMax;
 
 static double logIt( double v )
 {
